+++ rotate90.cpp
@@ -3,7 +3,7 @@
 
 void rotate90::process(const std::vector<std::complex<double> >& in, std::vector<std::complex<double> >& out) {
   out.resize(in.size());
-  for (int j = 0; j < in.size(); j++) {
+  for (size_t j = 0; j < in.size(); j++) {
     switch (phase) {
     case 0:        out[j] = in[j];        break;
     case 1:        out[j] = std::complex<double>(-std::imag(in[j]),std::real(in[j])); break;
